+++ CHANGELOG.md
@@ -38,6 +38,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 - Added HTTP renderer support for \Generator response body.
 - Added multibyte flag to LengthValidator to switch between strlen() and mb_strlen().
 - Added BetweenValidator construct parameter to allow for custom value validation. 
+- Added allowString flag to IntegerValidator to allow string representations of integers (e.g. in case of a URL value).
 
 ### Changed